feat(pragma_consumer): 🔨
