+++ pragma-node/src/servers/app/mod.rs
@@ -36,7 +36,7 @@ pub async fn run_app_server(config: &Config, state: AppState) {
             handlers::get_ohlc::get_ohlc,
             handlers::subscribe_to_entry::subscribe_to_entry,
             handlers::get_volatility::get_volatility,
-            handlers::onchain::get_onchain,
+            handlers::onchain::get_entry::get_onchain_entry,
             handlers::onchain::get_history::get_onchain_history,
             handlers::onchain::get_checkpoints::get_onchain_checkpoints,
             handlers::onchain::get_publishers::get_onchain_publishers,
@@ -58,8 +58,8 @@ pub async fn run_app_server(config: &Config, state: AppState) {
                 handlers::subscribe_to_entry::SubscribeToEntryResponse,
                 handlers::get_volatility::GetVolatilityResponse,
                 handlers::get_ohlc::GetOHLCResponse,
-                handlers::onchain::GetOnchainParams,
-                handlers::onchain::GetOnchainResponse,
+                handlers::onchain::get_entry::GetOnchainEntryParams,
+                handlers::onchain::get_entry::GetOnchainEntryResponse,
                 handlers::onchain::get_checkpoints::GetOnchainCheckpointsParams,
                 handlers::onchain::get_checkpoints::GetOnchainCheckpointsResponse,
                 handlers::onchain::get_publishers::GetOnchainPublishersParams,
@@ -77,7 +77,7 @@ pub async fn run_app_server(config: &Config, state: AppState) {
                 types::entries::Entry,
                 types::entries::PerpEntry,
                 types::entries::FutureEntry,
-                handlers::onchain::OnchainEntry,
+                handlers::onchain::get_entry::OnchainEntry,
                 handlers::onchain::get_checkpoints::Checkpoint,
                 handlers::onchain::get_publishers::Publisher,
                 handlers::onchain::get_publishers::PublisherEntry,
