+++ pragma-node/src/utils/conversion.rs
@@ -1,6 +1,7 @@
 use std::str::FromStr;
 
 use bigdecimal::BigDecimal;
+
 use pragma_entities::InfraError;
 
 pub fn convert_via_quote(
